+++ lib/lighthouseProvider/LighthouseDevice.dart
@@ -1,35 +1,72 @@
 import 'dart:async';
-import 'dart:math';
 
-import 'package:flutter/foundation.dart';
-import 'package:flutter_blue/flutter_blue.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:lighthouse_pm/lighthouseProvider/ble/DeviceIdentifier.dart';
 import 'package:rxdart/rxdart.dart';
 
 import 'LighthousePowerState.dart';
 
-///The bluetooth service that handles the power state of the device.
-final Guid _POWER_SERVICE = Guid('00001523-1212-efde-1523-785feabcd124');
+abstract class LighthouseDevice {
+  ///
+  /// Get the device's name.
+  ///
+  String get name;
 
-///The characteristic that handles the power state of the device.
-final Guid _POWER_CHARACTERISTIC = Guid('00001525-1212-efde-1523-785feabcd124');
+  ///
+  /// Get the identifier for the device.
+  ///
+  LHDeviceIdentifier get deviceIdentifier;
 
-/// A single lighthouse device. (This doesn't mean that it is a valid device.
-/// if used outside of the [LighthouseProvider] then you can be (almost) certain
-/// that it is a valid device).
-///
-/// Top get a device use the [LighthouseProvider]'s
-/// [LighthouseProvider.lighthouseDevices] stream to get a list of currently
-/// visible devices.
-class LighthouseDevice {
-  LighthouseDevice(BluetoothDevice device) : _device = device;
+  ///
+  /// Disconnect from the device and clean up the connection.
+  ///
+  Future disconnect() async {
+    if (_powerStateSubscription != null) {
+      await _powerStateSubscription.cancel();
+    }
+    await this._powerState.close();
+    await internalDisconnect();
+  }
 
-  ///Get the name of the device.
-  String get name => _device.name;
+  ///
+  /// Disconnecting for the devices up in the chain.
+  ///
+  @protected
+  Future internalDisconnect();
 
-  ///Get the id (MAC) of the device.
-  DeviceIdentifier get id => _device.id;
+  ///
+  /// Convert a device specific state byte to a global `LighthousePowerState`.
+  ///
+  @protected
+  LighthousePowerState internalGetPowerStateFromByte(int byte);
 
-  ///Get the power state of the device as an int.
+  ///
+  /// Get the update interval that this device supports.
+  ///
+  @protected
+  Duration getUpdateInterval() {
+    return const Duration(milliseconds: 1000);
+  }
+
+  ///
+  /// Get the current power state as a device specific byte.
+  ///
+  @protected
+  Future<int /*?*/ > getCurrentState();
+
+  ///
+  /// Change the actual state of the device.
+  /// For implementing you don't have to check the [newState] to see if it's legal
+  /// since the parent [changeState] function already did this.
+  @protected
+  Future internalChangeState(LighthousePowerState newState);
+
+  ///
+  /// If the device has an open connection.
+  ///
+  bool get hasOpenConnection;
+
+  /// Get the power state of the device as a device specific int.
   Stream<int> get powerState {
     this._startPowerStateStream();
     return this._powerState.stream;
@@ -37,52 +74,12 @@ class LighthouseDevice {
 
   ///Get the power state of the device as a [LighthousePowerState] "enum".
   Stream<LighthousePowerState> get powerStateEnum => this.powerState.map((e) {
-        return LighthousePowerState.fromByte(e);
+        return internalGetPowerStateFromByte(e);
       });
 
-  final BluetoothDevice _device;
-  BluetoothCharacteristic _characteristic;
   BehaviorSubject<int> _powerState = BehaviorSubject.seeded(0xFF);
   StreamSubscription _powerStateSubscription;
-
-  ///Check if the device is a valid Lighthouse device.
-  ///
-  /// This method is already done by [LighthouseProvider] and thus doesn't
-  /// have to be done again.
-  ///
-  /// **Note:** This will also connect to the device.
-  Future<bool> isValid() async {
-    debugPrint('Connecting to device: ${this._device.id.toString()}');
-    await this._device.connect(timeout: Duration(seconds: 10)).catchError((e) {
-      debugPrint(
-          'Connection timedout for device: ${this._device.id.toString()}');
-      return false;
-    }, test: (e) => e is TimeoutException);
-
-    debugPrint('Finding service for device: ${this._device.id.toString()}');
-    List<BluetoothService> services = await this._device.discoverServices();
-    for (final service in services) {
-      if (service.uuid != _POWER_SERVICE) {
-        continue;
-      }
-      // Find the correct characteristic.
-      for (final characteristic in service.characteristics) {
-        if (characteristic.uuid == _POWER_CHARACTERISTIC) {
-          this._characteristic = characteristic;
-          return true;
-        }
-      }
-    }
-    return false;
-  }
-
-  ///Disconnect from the device.
-  Future disconnect() async {
-    debugPrint('Disconnecting from the powerstate');
-    await this._powerStateSubscription.cancel();
-    this._characteristic = null;
-    await this._device.disconnect();
-  }
+  bool _powerStateTransaction = false;
 
   ///Change the state of the device.
   ///
@@ -103,15 +100,10 @@ class LighthouseDevice {
       debugPrint('Cannot change powerstate to booting');
       return;
     }
-    if (this._characteristic == null) {
-      return;
-    }
-    await this._characteristic.write([newState.setByte], withoutResponse: true);
+    await internalChangeState(newState);
   }
 
-  bool _powerStateTransaction = false;
-
-  ///Start the power state stream.
+  /// Start the power state stream.
   ///
   /// If this method is called while there is already an active stream then it
   /// will do nothing.
@@ -124,18 +116,22 @@ class LighthouseDevice {
       return;
     }
     _powerStateTransaction = false;
-    _powerStateSubscription =
-        Stream.periodic(Duration(milliseconds: 1000)).listen((_) {
-      if (this._characteristic != null) {
+    _powerStateSubscription = Stream.periodic(getUpdateInterval()).listen((_) {
+      if (hasOpenConnection) {
         if (!_powerStateTransaction) {
-          _powerStateTransaction = true;
-          this._characteristic.read().then((data) {
-            if (data.length >= 1) {
-              this._powerState.add(data[0]);
+          getCurrentState().then((data) {
+            if (this._powerState.isClosed) {
+              this.disconnect();
+              return;
             }
+            if (data != null) {
+              this._powerState.add(data);
+            }
+            // TODO: handle `null` data.
             _powerStateTransaction = false;
-          }).catchError((error) {
-            debugPrint(error);
+          }).catchError((error, s) {
+            debugPrint(error.toString());
+            debugPrint(s.toString());
           });
         }
       } else {
@@ -147,78 +143,8 @@ class LighthouseDevice {
       }
     });
     _powerStateSubscription.onDone(() {
-      debugPrint('Cleaning-up powerstate subscription!');
+      debugPrint('Cleaning-up power state subscription!');
       _powerStateSubscription = null;
     });
-  }
-}
-
-/// This is a fake [LighthouseDevice] that won't actually do anything.
-///
-/// This always uses fake data and doesn't actually connect to a device. And
-/// always uses fake data.
-class LighthouseDeviceFake extends LighthouseDevice {
-  LighthouseDeviceFake() : super(null) {
-    final random = new Random();
-    for (var i = 0; i < 8; i++) {
-      this._name += random.nextInt(16).toRadixString(16).toUpperCase();
-    }
-    String id = '00:00:00:00:00:';
-    id += random.nextInt(16).toRadixString(16).toUpperCase();
-    id += random.nextInt(16).toRadixString(16).toUpperCase();
-    this._id = new DeviceIdentifier(id);
-
-    if (random.nextBool()) {
-      this._powerState.add(LighthousePowerState.ON.stateByte);
-    } else {
-      this._powerState.add(LighthousePowerState.STANDBY.stateByte);
-    }
-  }
-
-  String _name = 'LHB-';
-
-  @override
-  String get name => _name;
-
-  DeviceIdentifier _id;
-
-  @override
-  DeviceIdentifier get id => _id;
-
-  BehaviorSubject<int> _powerState = BehaviorSubject.seeded(0xFF);
-
-  @override
-  Stream<int> get powerState => _powerState.stream;
-
-  @override
-  Stream<LighthousePowerState> get powerStateEnum =>
-      this.powerState.map((e) => LighthousePowerState.fromByte(e));
-
-  @override
-  Future<bool> isValid() async {
-    return true;
-  }
-
-  @override
-  Future disconnect() async {
-    return;
-  }
-
-  @override
-  Future changeState(LighthousePowerState newState) async {
-    switch (newState) {
-      case LighthousePowerState.BOOTING:
-      case LighthousePowerState.UNKNOWN:
-        return;
-      case LighthousePowerState.ON:
-        this._powerState.add(LighthousePowerState.ON.setByte);
-        await Future.delayed(new Duration(milliseconds: 10));
-        this._powerState.add(LighthousePowerState.BOOTING.stateByte);
-        await Future.delayed(new Duration(milliseconds: 1200));
-        this._powerState.add(LighthousePowerState.ON.stateByte);
-        break;
-      case LighthousePowerState.STANDBY:
-        this._powerState.add(LighthousePowerState.STANDBY.setByte);
-    }
   }
 }