+++ lib/lighthouseProvider/LighthouseProvider.dart
@@ -1,14 +1,12 @@
 import 'dart:async';
-import 'dart:math';
 
 import 'package:flutter/foundation.dart';
-import 'package:flutter/foundation.dart' as Foundation;
 import 'package:flutter_blue/flutter_blue.dart';
-import 'package:lighthouse_pm/lighthouseProvider/LighthouseDevice.dart';
+import 'package:lighthouse_pm/lighthouseProvider/ble/DeviceIdentifier.dart';
+import 'package:lighthouse_pm/lighthouseProvider/deviceProviders/BLEDeviceProvider.dart';
 import 'package:rxdart/rxdart.dart';
-import 'package:rxdart/subjects.dart';
-import 'package:rxdart/transformers.dart';
 
+import 'LighthouseDevice.dart';
 import 'timeout/TimeoutContainer.dart';
 
 ///A provider for getting all [LighthouseDevice]s in the region.
@@ -38,46 +36,57 @@ class LighthouseProvider {
   /// [LighthouseDevice]s.
   ///
   Stream<List<LighthouseDevice>> get lighthouseDevices =>
-      _lightHouseDevices.stream
-          .map((devices) => devices.map((device) => device.data).toList());
+      _lightHouseDevices.stream.map((containers) =>
+          containers.map((container) => container.data).toList());
 
   /// Get an instance of [LighthouseProvider].
   static LighthouseProvider get instance {
-    if (Foundation.kReleaseMode) {
-      return LighthouseProvider._instance;
-    } else {
-      // easily change to use the actual provider or the fake one.
-      return LighthouseProvider._instance;
-      // return LighthouseProviderFake.instance;
-    }
+    return LighthouseProvider._instance;
   }
 
   static final LighthouseProvider _instance = LighthouseProvider._();
-  Set<DeviceIdentifier> _connectingDevices = Set();
+  Set<LHDeviceIdentifier> _connectingDevices = Set();
+  Set<LHDeviceIdentifier> _rejectedDevices = Set();
   BehaviorSubject<List<TimeoutContainer<LighthouseDevice>>> _lightHouseDevices =
       BehaviorSubject.seeded([]);
+  StreamSubscription _scanResultSubscription;
+  Set<BLEDeviceProvider> _bleDeviceProviders = Set();
+
+  void addBLEDeviceProvider(BLEDeviceProvider bleDeviceProvider) {
+    _bleDeviceProviders.add(bleDeviceProvider);
+  }
+
+  void removeBLEDeviceProvider(BLEDeviceProvider bleDeviceProvider) {
+    _bleDeviceProviders.remove(bleDeviceProvider);
+  }
 
   /// Start scanning for [LighthouseDevice]s.
   ///
   /// This will clear the old "known" [LighthouseDevice]s and start searching
   /// for new ones.
   ///
+  /// Will call the [cleanUp] function before starting the scan.
   /// Will call the [FlutterBlue.startScan] function in the background.
   Future startScan({
     ScanMode scanMode = ScanMode.lowLatency,
     Duration timeout,
   }) async {
     await cleanUp();
-    _startListeningScanResults();
+    await _startListeningScanResults();
     await FlutterBlue.instance
         .startScan(scanMode: scanMode, timeout: timeout, allowDuplicates: true);
   }
 
+  ///
+  /// Clean up any open connections that may still be left.
+  /// Will also clear out the current devices list.
+  ///
   Future cleanUp() async {
     await stopScan();
     await _disconnectOpenDevices();
     _lightHouseDevices.add(List());
     _connectingDevices.clear();
+    _rejectedDevices.clear();
   }
 
   /// Stop scanning for [LighthouseDevice]s.
@@ -96,10 +105,10 @@ class LighthouseProvider {
   ///
   /// Wil return the index of the device in the [_lightHouseDevices.value]
   /// [List] or `-1` if not found.
-  int _foundDeviceIndex(ScanResult device) {
+  int _foundDeviceIndex(LHDeviceIdentifier deviceIdentifier) {
     int index = 0;
     for (final activeDevice in this._lightHouseDevices.value) {
-      if (device.device.id == activeDevice.data.id) {
+      if (deviceIdentifier == activeDevice.data.deviceIdentifier) {
         return index;
       }
       index++;
@@ -109,73 +118,77 @@ class LighthouseProvider {
 
   /// Disconnect from all known and open devices.
   Future _disconnectOpenDevices() async {
+    for (final bleDeviceProvider in _bleDeviceProviders) {
+      await bleDeviceProvider.disconnectRunningDiscoveries();
+    }
     final list = this._lightHouseDevices.value;
     for (final device in list) {
       await device.data.disconnect();
     }
   }
 
-  StreamSubscription _scanResultSubscription;
-
   /// Start the stream for listening to the [LighthouseDevice]s.
-  ///
-  /// This will try and re-use a previous active stream if one already exists.
-  void _startListeningScanResults() {
+  Future _startListeningScanResults() async {
     if (_scanResultSubscription != null) {
-      if (_scanResultSubscription.isPaused) {
-        _scanResultSubscription.resume();
-        return;
+      if (!_scanResultSubscription.isPaused) {
+        _scanResultSubscription.pause();
       }
-      _scanResultSubscription.cancel();
+      await _scanResultSubscription.cancel();
+      _scanResultSubscription = null;
     }
 
     _scanResultSubscription = FlutterBlue.instance.scanResults
-        .map((devices) {
-          // Filter out all devices who's name doesn't start with `LHB-`.
+        .map((scanResults) {
+          // Filter out all devices that don't have a correct name.
           final List<ScanResult> output = List();
-          for (final device in devices) {
-            if (device.device.name.startsWith('LHB-')) {
-              output.add(device);
-            } else {
-              //debugPrint(
-              //    'Filtered out device that didn\'t have the correct name. Mac: ${device.device.id.toString()}');
+          for (final scanResult in scanResults) {
+            for (final bleDeviceProviders in _bleDeviceProviders) {
+              if (bleDeviceProviders.nameCheck(scanResult.device.name)) {
+                output.add(scanResult);
+                break;
+              }
             }
           }
           return output;
         })
         // Give the listener at least 2ms to process the data before firing again.
         .debounce((_) => TimerStream(true, Duration(milliseconds: 2)))
-        .listen((devices) {
-          for (final device in devices) {
-            if (this._connectingDevices.contains(device.device.id)) {
-              debugPrint(
-                  'Found a device that is already being scanned. Mac: ${device.device.id.toString()}');
+        .listen((scanResults) {
+          if (scanResults.isEmpty) {
+            return;
+          }
+          for (final scanResult in scanResults) {
+            if (this._connectingDevices.contains(
+                LHDeviceIdentifier.fromFlutterBlue(scanResult.device.id))) {
               continue;
             }
-            final index = this._foundDeviceIndex(device);
+            if (this._rejectedDevices.contains(
+                LHDeviceIdentifier.fromFlutterBlue(scanResult.device.id))) {
+              continue;
+            }
+            // Update the last seen item.
+            final index = this._foundDeviceIndex(
+                LHDeviceIdentifier.fromFlutterBlue(scanResult.device.id));
             if (index >= 0) {
-              debugPrint(
-                  'Found a device that was already known. Mac: ${device.device.id.toString()}');
               this._lightHouseDevices.value[index].lastSeen = DateTime.now();
               continue;
             }
-
-            // A new [LighthouseDevice], let's make sure it's valid.
-            final lighthouseDevice = LighthouseDevice(device.device);
-            this._connectingDevices.add(device.device.id);
-            lighthouseDevice.isValid().then((valid) {
-              if (valid) {
+            // Possibly a new lighthouse, let's make sure it's valid.
+            this
+                ._connectingDevices
+                .add(LHDeviceIdentifier.fromFlutterBlue(scanResult.device.id));
+            _getLighthouseDevice(scanResult.device).then((lighthouseDevice) {
+              if (lighthouseDevice == null) {
                 debugPrint(
-                    'Found a new lighthouse! Mac: ${device.device.id.toString()}');
-                // Add the new device to the list and publish it to the stream.
+                    'Found a non valid device! Mac: ${scanResult.device.id.toString()}');
+                this._rejectedDevices.add(
+                    LHDeviceIdentifier.fromFlutterBlue(scanResult.device.id));
+              } else {
                 final list = this._lightHouseDevices.value;
                 list.add(TimeoutContainer<LighthouseDevice>(lighthouseDevice));
                 this._lightHouseDevices.add(list);
-              } else {
-                debugPrint(
-                    'Found a non valid device! Mac: ${device.device.id.toString()}');
               }
-              this._connectingDevices.remove(device.device.id);
+              this._connectingDevices.remove(scanResult.device.id);
             });
           }
         });
@@ -184,47 +197,21 @@ class LighthouseProvider {
       this._scanResultSubscription = null;
     });
   }
-}
 
-/// A [LighthouseProvider] that always returns fake data and doesn't actually
-/// communicate with any bluetooth service.
-///
-/// This is only used for testing and creating screenshots.
-class LighthouseProviderFake extends LighthouseProvider {
-  LighthouseProviderFake._() : super._();
-
-  @override
-  Future stopScan() async {
-    return;
-  }
-
-  @override
-  Future startScan(
-      {ScanMode scanMode = ScanMode.lowLatency, Duration timeout}) async {
-    this._lighthouseDevicesFake.add(List());
-    final list = this._lighthouseDevicesFake.value;
-    final random = new Random();
-    for (var i = 0; i < 20; i++) {
-      await Future.delayed(
-          new Duration(milliseconds: random.nextInt(500) + 200));
-      list.add(new LighthouseDeviceFake());
-      this._lighthouseDevicesFake.add(list);
+  ///
+  /// Will return `null` if no device provider could validate the device.
+  Future<LighthouseDevice> _getLighthouseDevice(BluetoothDevice device) async {
+    debugPrint('Trying to connect to device with name: ${device.name}');
+    for (final bLEDeviceProvider in _bleDeviceProviders) {
+      if (!bLEDeviceProvider.nameCheck(device.name)) {
+        continue;
+      }
+      final LighthouseDevice lighthouseDevice =
+          await bLEDeviceProvider.getDevice(device);
+      if (lighthouseDevice != null) {
+        return lighthouseDevice;
+      }
     }
-  }
-
-  BehaviorSubject<List<LighthouseDevice>> _lighthouseDevicesFake =
-      BehaviorSubject.seeded([]);
-
-  @override
-  Stream<List<LighthouseDevice>> get lighthouseDevices =>
-      _lighthouseDevicesFake.stream;
-
-  static LighthouseProviderFake _instance;
-
-  static LighthouseProviderFake get instance {
-    if (LighthouseProviderFake._instance == null) {
-      LighthouseProviderFake._instance = LighthouseProviderFake._();
-    }
-    return LighthouseProviderFake._instance;
+    return null;
   }
 }